Error Messages

+++ src/sender_udp.h
@@ -1,7 +1,7 @@
 #ifndef SENDER_UDP_H_
 #define SENDER_UDP_H_
 
-void prepareHeader(char *buffer, unsigned short nlength, char *name, unsigned long filelength);
+void prepareHeader(char *buffer, unsigned short nlength, char *name, unsigned int filelength);
 
 //char *getSha1(char *buff, int bufferlength);
